+++ public/templates/default/html/global/nav.php
@@ -73,8 +73,8 @@
                         </a>
                         <div class="dropdown-menu p-0 dropdown-menu-right border-0 font-size-sm" style="z-index: 1065;">
                             <div class="text-center d-block py-2 aw-nav aw-dropdown-nav text-center aw-answer-sort" style="min-width: 100px">
-                                <a href="{:frelink_publish_url('question')}" class="py-1 dropdown-item">{:L('FAQ 条目')}</a>
-                                <a href="{:frelink_publish_url('article')}" class="py-1 dropdown-item">{:L('综述 / 观察')}</a>
+                                <a href="{:frelink_publish_url('question')}" class="py-1 dropdown-item">{:L('FAQ')}</a>
+                                <a href="{:frelink_publish_url('article')}" class="py-1 dropdown-item">{:L('知识内容')}</a>
                                 {volist name=":config('aws.publish')" id="v"}
                                 {if isset($v['url']) && $v['url']}
                                 <a href="{:url($v['url'])}" class="py-1 dropdown-item">{$v.title}</a>